Add files via upload
This project demonstrates a way to create an ERC20 token, installation it the usage of Remix IDE, and engage with it thru a web interface. The smart agreement permits the agreement owner to mint tokens to a specified deal with, and any user can transfer or burn tokens. This mission combines abilities in Solidity programming, blockchain interaction, and net improvement.